Inline RenderingSystem into GameWorld
